refactor(accounts): rename companies-grid modules to accounts-grid
P1-6 slice 2: file and symbol rename only; DB companies table unchanged.
