+++ docs/arbeitspaket-perf-5-bundle-tti.md
@@ -14,7 +14,7 @@
 
 - **170** `'use client'`-Dateien, **0** `next/dynamic` (kein Code-Splitting), **kein** Bundle-Analyzer.
 - **Schwere Libs — korrigiert:** `recharts` ist **nicht** im Projekt (meine frühere Angabe war falsch). `@react-pdf/renderer` ist **bereits server-only** (`app/api/pdf/*`, `lib/evidence/pdf/*`) — kein Client-Leak. `framer-motion` ist in `package.json`, hat aber **keine Imports** → Entfernungs-Kandidat. **Echter Client-Leak: nur `xlsx`** via `lib/accounts/companies-import-template.ts` (`import * as XLSX`), gezogen über `companies-import-dialog.tsx`.
-- 12 Client-Komponenten > 500 Z. (u. a. `dashboard-overview` 1457, `deal-desk-client` 1135, `companies-grid` 830, `reference-detail-sheet` 636, `account-detail-nda-popover` 623); **34** Dialog/Sheet/Modal-Komponenten.
+- 12 Client-Komponenten > 500 Z. (u. a. `dashboard-overview` 1457, `deal-desk-client` 1135, `accounts-grid` 830, `reference-detail-sheet` 636, `account-detail-nda-popover` 623); **34** Dialog/Sheet/Modal-Komponenten.
 - **T1 (Analyzer) ist die Wahrheit** — die obigen Verdächtigen sind Hypothesen; der Build-Report priorisiert verbindlich.
 
 > **Nach T1: klare Gewinne nehmen, dann stoppen.** Sichere Wins: `xlsx` → Server/API, `framer-motion` entfernen, die schwersten 2–3 Dialoge der Hot-Routen lazy. **Nicht** reflexartig alle 34 Dialoge umstellen, wenn der T1-Report nur marginalen Rest-Gewinn zeigt — Diminishing Returns.
